pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "plnn"
dynamic = ["version"]
requires-python = ">= 3.9"
readme = "README.md"
license = { file = "LICENSE" }
description = "A machine learning approach to parameterized landscapes"
authors = [
  {name = "Addison Howe"},
]
dependencies = [
  "jax>=0.4",
  "numpy==1.26",
  "matplotlib==3.8",
  "scikit-learn==1.5",
  "torch==2.0",
  "torchvision",
  "equinox>=0.11",
  "optax>=0.1",
  "pyyaml==6.0",
  "diffrax>=0.6",
  "pytest==7.4",
  "tqdm",
  "ipykernel",
]

[project.scripts]
plnn = "plnn.__main__:main"
model_eval = "plnn.eval.__main__:main"
generate_data = "plnn.data_generation.__main__:main"

[project.optional-dependencies]
cuda = [
  "cuda-compat==12.4",
  "numpy==1.25",
  "torch==2.0",
  "jax[cuda12_pip]",
  "torchvision",
  "torchaudio",
]
dev = [
  "setuptools-scm",
]

[tool.setuptools_scm]
version_scheme = "post-release"
local_scheme = "node-and-date"
write_to = "src/plnn/_version.py"
